@@ -1,15 +1,13 @@
-module github.com/Shopify/goose
+module github.com/Shopify/goose/v2
 
-go 1.15
+go 1.20
 
 require (
 	github.com/DataDog/datadog-go v4.8.1+incompatible
-	github.com/Microsoft/go-winio v0.5.0 // indirect
 	github.com/bugsnag/bugsnag-go/v2 v2.1.1
 	github.com/bugsnag/panicwrap v1.3.4
 	github.com/google/pprof v0.0.0-20210804190019-f964ff605595
 	github.com/gorilla/mux v1.8.0
-	github.com/ianlancetaylor/demangle v0.0.0-20210724235854-665d3a6fe486 // indirect
 	github.com/imdario/mergo v0.3.12
 	github.com/leononame/clock v0.1.6
 	github.com/nu7hatch/gouuid v0.0.0-20131221200532-179d4d0c4d8d
@@ -18,9 +16,20 @@ require (
 	github.com/stretchr/testify v1.8.1
 	golang.org/x/net v0.0.0-20210813160813-60bc85c4be6d
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
+	gopkg.in/tomb.v2 v2.0.0-20161208151619-d5d1b5820637
+)
+
+require (
+	github.com/Microsoft/go-winio v0.5.0 // indirect
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/gofrs/uuid v4.0.0+incompatible // indirect
+	github.com/ianlancetaylor/demangle v0.0.0-20210724235854-665d3a6fe486 // indirect
+	github.com/kardianos/osext v0.0.0-20190222173326-2bc1f35cddc0 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/objx v0.5.0 // indirect
 	golang.org/x/sys v0.0.0-20210817190340-bfb29a6856f2 // indirect
 	golang.org/x/text v0.3.7 // indirect
 	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
-	gopkg.in/tomb.v2 v2.0.0-20161208151619-d5d1b5820637
 	gopkg.in/yaml.v2 v2.4.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
