@@ -5,7 +5,7 @@ import (
 	"bytes"
 	"context"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net"
 	"net/http"
 	"net/http/httptest"
@@ -17,9 +17,9 @@ import (
 	"github.com/stretchr/testify/assert"
 	"gopkg.in/tomb.v2"
 
-	"github.com/Shopify/goose/metrics"
-	"github.com/Shopify/goose/safely"
-	"github.com/Shopify/goose/statsd"
+	"github.com/Shopify/goose/v2/metrics"
+	"github.com/Shopify/goose/v2/safely"
+	"github.com/Shopify/goose/v2/statsd"
 )
 
 func TestRequestMetricsMiddleware(t *testing.T) {
@@ -70,7 +70,7 @@ func TestRequestMetricsMiddleware(t *testing.T) {
 	res, err := http.DefaultClient.Do(req)
 	assert.NoError(t, err)
 	assert.Equal(t, http.StatusOK, res.StatusCode)
-	body, err := ioutil.ReadAll(res.Body)
+	body, err := io.ReadAll(res.Body)
 	assert.NoError(t, err)
 	assert.Equal(t, "hello world", string(body))
 
