@@ -6,7 +6,7 @@ import (
 
 	"golang.org/x/sync/semaphore"
 
-	"github.com/Shopify/goose/statsd"
+	"github.com/Shopify/goose/v2/statsd"
 )
 
 const (
