@@ -14,10 +14,10 @@ import (
 
 	"gopkg.in/tomb.v2"
 
-	"github.com/Shopify/goose/logger"
-	"github.com/Shopify/goose/metrics"
-	"github.com/Shopify/goose/safely"
-	"github.com/Shopify/goose/statsd"
+	"github.com/Shopify/goose/v2/logger"
+	"github.com/Shopify/goose/v2/metrics"
+	"github.com/Shopify/goose/v2/safely"
+	"github.com/Shopify/goose/v2/statsd"
 )
 
 var log = logger.New("genmain")
