@@ -8,8 +8,8 @@ import (
 
 	"github.com/sirupsen/logrus"
 
-	"github.com/Shopify/goose/logger"
-	"github.com/Shopify/goose/statsd"
+	"github.com/Shopify/goose/v2/logger"
+	"github.com/Shopify/goose/v2/statsd"
 )
 
 var log = logger.New("shell")
