@@ -5,14 +5,14 @@ import (
 	"fmt"
 	"sort"
 
-	"github.com/Shopify/goose/logger"
+	"github.com/Shopify/goose/v2/logger"
 )
 
 // This create a private key-space in the Context, meaning that only this package can get or set "contextKey" types
-type contextKey int
+type contextKey struct{}
 
-const (
-	tagsKey contextKey = iota
+var (
+	tagsKey = contextKey{}
 )
 
 type Tags map[string]interface{}
